Rename package antlr4 to antlr
