+++ runtime/Go/antlr/ATNConfigSet.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import (
 	"fmt"
