+++ runtime/Go/antlr/DFA.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import "sort"
