+++ runtime/Go/antlr/ErrorStrategy.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import (
 	"fmt"
