+++ runtime/Go/antlr/FileStream.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import (
 	"bytes"
