+++ runtime/Go/antlr/IntStream.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 type IntStream interface {
 	Consume()
