+++ runtime/Go/antlr/LL1Analyzer.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import (
 	"fmt"
