+++ runtime/Go/antlr/LexerAction.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import "strconv"
