+++ runtime/Go/antlr/LexerActionExecutor.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 // Represents an executor for a sequence of lexer actions which traversed during
 // the Matching operation of a lexer rule (token).
