+++ runtime/Go/antlr/ParserRuleContext.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import (
 	"reflect"
