+++ runtime/Go/antlr/RuleContext.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 //  A rule context is a record of a single rule invocation. It knows
 //  which context invoked it, if any. If there is no parent context, then
