+++ runtime/Go/antlr/TokenSource.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 type TokenSource interface {
 	NextToken() Token
