+++ runtime/Go/antlr/Tree.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 // The basic notion of a tree has a parent, a payload, and a list of children.
 //  It is the most abstract interface for all the trees used by ANTLR.
