+++ runtime/Go/antlr/Utils.go
@@ -1,4 +1,4 @@
-package antlr4
+package antlr
 
 import (
 	"errors"
